requirements.txt
gitpython>=3.1.0
jinja2>=3.0.0
pyyaml>=6.0
requests>=2.25.0
pathlib2>=2.3.0; python_version < '3.4'